Implement base definition reviewer (#1)
* Add BaseDefinitionReviewer

BaseDefinitionReviewer checks the types of the two StructureDefinitions.
If they match, the result is equivalent. Otherwise, the result is
disjoint. Type is required on StructureDefinition, so there is no check
for missing types.

* Check for missing type

Type is required on StructureDefinition, so if it is not present, show
the user a strongly worded message indicating that the current situation
is not okay.