Review extension elements with ExtensionReviewer
Extension elements on the same path, but potentially with different ids,
are compared by the ExtensionReviewer. The profiles available on
corresponding elements are checked. If the same profile is available on
both StructureDefinitions, the cardinalities are compared. Otherwise, a
missing profile is sufficient to determine the result. If no profiles
are present, no comparison is performed.