Always include path on ExtensionReviewer results
When creating a Review for an extension element that is only present on
one of the StructureDefinitions being reviewed, set the path for the
other StructureDefinition to the path (not id) of the element being
reviewed.

Numeric range comparison logic moved to utility function. This will
probably only be used for cardinality range comparisons, but other
scenarios might arise, I guess?